Rename fake-provider-engine to fake-provider
